Move transform function to /services
